refactor: apply namespaces fix
