--- Source/Application/Kysect.Shreks.Application.Commands/Parser/IShreksCommandParser.cs
@@ -1,6 +1,0 @@
-namespace Kysect.Shreks.Application.Commands.Commands;
-
-public interface IShreksCommandParser
-{
-    IShreksCommand Parse(string commandStr);
-}
