+++ Source/Application/Kysect.Shreks.Application.Commands/Parsers/IShreksCommandParser.cs
@@ -1,0 +1,8 @@
+using Kysect.Shreks.Application.Commands.Commands;
+
+namespace Kysect.Shreks.Application.Commands.Parsers;
+
+public interface IShreksCommandParser
+{
+    IShreksCommand Parse(string commandStr);
+}
