+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github/Processors/ShreksWebhookEventProcessor.cs
@@ -1,5 +1,6 @@
 using Kysect.Shreks.Application.Abstractions.DataAccess;
 using Kysect.Shreks.Application.Commands.Commands;
+using Kysect.Shreks.Application.Commands.Parsers;
 using Kysect.Shreks.Integration.Github.ContextFactory;
 using Kysect.Shreks.Integration.Github.Entities;
 using MediatR;
