+++ Source/Infrastructure/Kysect.Shreks.Seeding/EntityGenerators/DeadlinePolicy/DeadlinePolicyGenerator.cs
@@ -1,6 +1,5 @@
 using Bogus;
 using Kysect.Shreks.Core.DeadlinePolicies;
-using Kysect.Shreks.Core.ValueObject;
 using Kysect.Shreks.Seeding.Extensions;
 using Kysect.Shreks.Seeding.Options;
